moving helper function to header and make them inline
